+++ scheduler/src/cook/scheduler/constraints.clj
@@ -22,6 +22,7 @@
             [cook.cached-queries :as cached-queries]
             [cook.compute-cluster :as cc]
             [cook.config :as config]
+            [cook.config-incremental :as config-incremental]
             [cook.group :as group]
             [cook.regexp-tools :as regexp-tools]
             [cook.tools :as util]
@@ -250,13 +251,26 @@
     [this _ vm-attributes _]
     (job-constraint-evaluate this _ vm-attributes)))
 
+(defn- resolve-incremental-default-job-constraints
+  "Resolve any incremental configurations used for default job constraint patterns"
+  [default-job-constraints {:keys [job/uuid]}]
+  (->> default-job-constraints
+       (map (fn [{:keys [constraint/pattern constraint/pattern-fallback] :as constraint}]
+              (if (string? pattern)
+                constraint
+                (let [[resolved-pattern _] (config-incremental/resolve-incremental-config uuid pattern pattern-fallback)]
+                  (-> constraint
+                    (dissoc :constraint/pattern-fallback)
+                    (assoc :constraint/pattern resolved-pattern))))))))
+
 (defn job->default-constraints
   "Returns the list of default constraints configured for the job's pool"
   [job]
-  (regexp-tools/match-based-on-pool-name
-    (config/default-job-constraints)
-    (cached-queries/job->pool-name job)
-    :default-constraints))
+  (resolve-incremental-default-job-constraints
+    (regexp-tools/match-based-on-pool-name
+      (config/default-job-constraints)
+      (cached-queries/job->pool-name job)
+      :default-constraints) job))
 
 (def machine-type-constraint-attributes
   #{"cpu-architecture" "node-family" "node-type"})